kejani_backend/requirements.txt
iniconfig==2.3.0
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
python-decouple==3.8
whitenoise==6.11.0
